Refactor: users,auth 디렉토리 auth 디렉토리로 병합
