+++ src/main/java/sportsmatchingservice/auth/controller/UserController.java
@@ -1,12 +1,12 @@
-package sportsmatchingservice.users.controller;
+package sportsmatchingservice.auth.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import sportsmatchingservice.auth.dto.UserSignupDto;
 import sportsmatchingservice.constant.ErrorCode;
 import sportsmatchingservice.constant.dto.ApiDataResponse;
-import sportsmatchingservice.users.dto.UserPostDto;
-import sportsmatchingservice.users.service.UserService;
+import sportsmatchingservice.auth.service.UserService;
 
 
 @RestController
@@ -16,9 +16,9 @@ public class UserController {
     private final UserService userService;
 
     @ResponseStatus(HttpStatus.CREATED)
-    @PostMapping("/signup")
-    public ApiDataResponse signup(@RequestBody UserPostDto userPostDto) {
-        boolean result = userService.createUser(userPostDto);
+    @PostMapping("")
+    public ApiDataResponse signup(@RequestBody UserSignupDto userSignupDto) {
+        boolean result = userService.createUser(userSignupDto);
         if (result == true) {
             return ApiDataResponse.of(ErrorCode.CREATED, result);
         } else {
