+++ src/main/java/sportsmatchingservice/auth/service/UserService.java
@@ -1,15 +1,15 @@
-package sportsmatchingservice.users.service;
+package sportsmatchingservice.auth.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import sportsmatchingservice.auth.dto.UserSignupDto;
 import sportsmatchingservice.auth.utils.CustomAuthorityUtils;
 import sportsmatchingservice.constant.ErrorCode;
 import sportsmatchingservice.exceptions.exception.DuplicatedDataException;
-import sportsmatchingservice.users.domain.User;
-import sportsmatchingservice.users.dto.UserPostDto;
-import sportsmatchingservice.users.repository.UserRepository;
+import sportsmatchingservice.auth.domain.User;
+import sportsmatchingservice.auth.repository.UserRepository;
 
 import java.util.List;
 
@@ -24,12 +24,12 @@ public class UserService {
     private final CustomAuthorityUtils authorityUtils;
 
 
-    public boolean createUser(UserPostDto userPostDto) {
+    public boolean createUser(UserSignupDto userSignupDto) {
         try {
-            if (userPostDto == null) return false;
+            if (userSignupDto == null) return false;
 
-            User user = userPostDto.toEntity();
-            user.setPassword(passwordEncoder.encode(userPostDto.getPassword()));
+            User user = userSignupDto.toEntity();
+            user.setPassword(passwordEncoder.encode(userSignupDto.getPassword()));
 
             List<String> roles = authorityUtils.createRoles(user.getEmail());
             user.setRoles(roles);
